+++ lib/IDE/CMakeLists.txt
@@ -12,6 +12,7 @@ add_swift_host_library(swiftIDE STATIC
   CodeCompletionResultPrinter.cpp
   CodeCompletionResultType.cpp
   CodeCompletionString.cpp
+  CodeCompletionStringBuilder.cpp
   CodeCompletionStringPrinter.cpp
   CommentConversion.cpp
   CompletionLookup.cpp
